feat: Add query parameters to getPackages function
This commit adds query parameters to the getPackages function in usePackage.ts, allowing for more specific package retrieval. The queryParams object is now used in the API request to filter packages based on the selected type. This change enhances the functionality of the application by providing more flexibility in package retrieval.