feat: Update start script to use Vite preview
This commit updates the start script in the package.json file to use Vite's preview command. By using the preview command, the development server will start and the application will be accessible at the specified port. This change improves the development workflow by providing a convenient way to preview the application during development.